CARDS-1984: New APIs for sending emails Bugfix: sometimes the email templates fail to load
