+++ modules/patient-portal/src/main/java/io/uhndata/cards/patients/emailnotifications/AppointmentEmailNotificationsFactory.java
@@ -16,16 +16,7 @@
  */
 package io.uhndata.cards.patients.emailnotifications;
 
-import java.io.IOException;
-import java.util.Map;
-
-import javax.jcr.Node;
-import javax.jcr.RepositoryException;
-import javax.jcr.Session;
-
 import org.apache.commons.lang3.StringUtils;
-import org.apache.sling.api.resource.LoginException;
-import org.apache.sling.api.resource.ResourceResolver;
 import org.apache.sling.api.resource.ResourceResolverFactory;
 import org.apache.sling.commons.messaging.mail.MailService;
 import org.apache.sling.commons.scheduler.ScheduleOptions;
@@ -42,7 +33,6 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import io.uhndata.cards.auth.token.TokenManager;
-import io.uhndata.cards.emailnotifications.EmailTemplate;
 import io.uhndata.cards.forms.api.FormUtils;
 import io.uhndata.cards.metrics.Metrics;
 import io.uhndata.cards.patients.api.PatientAccessConfiguration;
@@ -124,8 +114,7 @@ public final class AppointmentEmailNotificationsFactory
         // Instantiate the Runnable
         final Runnable notificationsJob = new GeneralNotificationsTask(this.resolverFactory, this.resolverProvider,
             this.tokenManager, this.mailService, this.formUtils, this.patientAccessConfiguration, config.name(),
-            config.clinicId(),
-            buildTemplate(config.emailConfiguration()), config.daysToVisit());
+            config.clinicId(), config.emailConfiguration(), config.daysToVisit());
 
         try {
             this.scheduler.schedule(notificationsJob, notificationsOptions);
@@ -149,38 +138,5 @@ public final class AppointmentEmailNotificationsFactory
         } else {
             LOGGER.error("Failed to unschedule {}", jobName);
         }
-    }
-
-    private EmailTemplate buildTemplate(final String configurationNodePath)
-    {
-        try (ResourceResolver rr = this.resolverFactory
-            .getServiceResourceResolver(Map.of(ResourceResolverFactory.SUBSERVICE, "EmailNotifications"))) {
-            final Session session = rr.adaptTo(Session.class);
-            // Sometimes the notification is activated before the template node is imported
-            // If that is the case, wait a little bit and try again
-            for (int i = 0; i < 10; ++i) {
-                if (session.nodeExists(configurationNodePath)) {
-                    break;
-                } else {
-                    try {
-                        Thread.sleep(10_000);
-                    } catch (InterruptedException e) {
-                        // We're not expecting any interruptions
-                    }
-                }
-            }
-            if (!session.nodeExists(configurationNodePath)) {
-                throw new IllegalStateException(
-                    "Configured email template " + configurationNodePath + " was not found");
-            }
-            final Node configuration = session.getNode(configurationNodePath);
-            return EmailTemplate.builder(configuration, rr).build();
-        } catch (LoginException e) {
-            LOGGER.warn("Missing rights configuration for Email Notifications");
-        } catch (RepositoryException | IOException e) {
-            LOGGER.warn("Failed to read the email configuration: {}", e.getMessage(), e);
-            throw new IllegalStateException();
-        }
-        return null;
     }
 }
